Update Calculate_flow.py
New method for flow calculation: the estimated flow at the subcatchment level is proportionally distributed within the river sections within the subcatchment. No more CATCH_ID and no more dissolving. Dissolving messes up the geometry of the river sections. It then causes problems with the accumulation tool in the scenario where a WWTP is in a subcatchment with multiple river sections.